+++ release_notes.md
@@ -1,6 +1,16 @@
 Spider2 Changelog
 ================
 
+## Release version X.Y.Z
+*XXXX.XX.XX*
+
+### New Feature
+
+### Changes
+
+### Bug fix
+
+
 ## Release version 0.0.0
 *2020.04.14*