+++ release_notes.md
@@ -1,8 +1,8 @@
 Spider2 Changelog
 ================
 
-## Release version X.Y.Z
-*XXXX.XX.XX*
+## Release version 0.1.0
+*2020.06.18*
 
 ### New Feature